requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg[binary]==3.3.2
Flask-Mail==0.10.0
